Update libvirt-go to 7.0.0
The current version (5.1.0) does not compile on 32-bit architectures
such as the Raspberry Pi. Updating the dependency to the latest
version and tweaking build errors allows the unit tests to pass.

+++ libvirt/resource_libvirt_domain.go
@@ -460,7 +460,7 @@ func resourceLibvirtDomainCreate(d *schema.ResourceData, meta interface{}) error
 		Unit:  "MiB",
 	}
 	domainDef.VCPU = &libvirtxml.DomainVCPU{
-		Value: d.Get("vcpu").(int),
+		Value: uint(d.Get("vcpu").(int)),
 	}
 	domainDef.Description = d.Get("description").(string)
